chore: refactored exception handler.
