+++ src/routers/classification/documentation.py
@@ -6,7 +6,7 @@ from services.postgres.connection import get_db
 from src.schema.response import ResponseDefault
 from src.schema.request_format import AllowedIpAddress
 from utils.query.general import find_record
-from utils.custom_errors import AccessUnauthorized, ServiceError, DatabaseQueryError, DiVA
+from utils.custom_error import AccessUnauthorized, ServiceError, DatabaseQueryError, DiVA
 from services.postgres.models import (
     CategoryDataDocumentation,
     ObjectDocumentationDetails,
