+++ tests/unit_test/test_query_general_retrieve_all.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import AsyncMock, MagicMock, patch
 from sqlalchemy.ext.asyncio import AsyncSession
-from utils.custom_errors import DatabaseQueryError, DataNotFoundError
+from utils.custom_error import DatabaseQueryError, DataNotFoundError
 from fixtures.mock_table import example_table
