+++ utils/query/model_accuracy.py
@@ -1,7 +1,7 @@
 from sqlalchemy import insert
 from services.postgres.connection import database_connection
 from services.postgres.models import ModelAccuracy
-from utils.custom_errors import DatabaseQueryError
+from utils.custom_error import DatabaseQueryError
 from utils.logger import logging
 from utils.helper import local_time
